+++ src/main/java/com/esp/service/GenericService.java
@@ -1,5 +1,6 @@
 package com.esp.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -7,7 +8,7 @@ public interface GenericService<T> {
 	
 public void add(T t);
 
-public T fetch(int id); 
+public T fetch(BigDecimal id);
 
 public List<T> fetchByParam(Object obj);
